🎨 Improve use of options for the ldap command
